+++ RNAentropy/intl11.h
@@ -1,4 +1,4 @@
-PUBLIC int int11_37[NBPAIRS+1][NBPAIRS+1][5][5] =
+PUBLIC double int11_37[NBPAIRS+1][NBPAIRS+1][5][5] =
 {{{{   INF,   INF,   INF,   INF,   INF}
   ,{   INF,   INF,   INF,   INF,   INF}
   ,{   INF,   INF,   INF,   INF,   INF}
