+++ RNAentropy/intl21.h
@@ -1,4 +1,4 @@
-PUBLIC int int21_37[NBPAIRS+1][NBPAIRS+1][5][5][5] =
+PUBLIC double int21_37[NBPAIRS+1][NBPAIRS+1][5][5][5] =
 {{{{{   INF,   INF,   INF,   INF,   INF}
    ,{   INF,   INF,   INF,   INF,   INF}
    ,{   INF,   INF,   INF,   INF,   INF}
